Make code reusable in addClientModules
