+++ src/transforms/v2-to-v3/modules/importModule/index.ts
@@ -1,0 +1,2 @@
+export * from "./addClientDefaultModule";
+export * from "./addClientNamedModule";
