+++ src/transforms/v2-to-v3/modules/types.ts
@@ -6,11 +6,6 @@ export interface ClientModulesOptions {
   v3ClientPackageName: string;
 }
 
-export interface RequirePropertyOptions {
-  keyName: string;
-  valueName?: string;
-}
-
 export interface ImportSpecifierOptions {
   importedName: string;
   localName?: string;
